remove old priority & replace update with schedule step in priority
